+++ neoforge/src/main/java/com/jvn/villagerretaliation/config/VillagerRetaliationConfig.java
@@ -261,7 +261,13 @@ public final class VillagerRetaliationConfig {
     public static final ConfigValue<Boolean> ENABLE_VILLAGER_SLEEP_HEALING = bind("combat.enableVillagerSleepHealing", Boolean.class);
     public static final ConfigValue<Double> VILLAGER_SLEEP_HEALING_MAX_HEALTH_PERCENT = bind("combat.villagerSleepHealingMaxHealthPercent", Double.class);
     public static final ConfigValue<Boolean> ENABLE_VILLAGER_DOWNED_STATE = bind("combat.enableVillagerDownedState", Boolean.class);
+    public static final ConfigValue<Boolean> ALL_VILLAGERS_USE_DOWNED_STATE = bind("combat.allVillagersUseDownedState", Boolean.class);
+    public static final ConfigValue<Boolean> RAID_VILLAGERS_USE_DOWNED_STATE = bind("combat.raidVillagersUseDownedState", Boolean.class);
+    public static final ConfigValue<Boolean> HIRED_VILLAGERS_USE_DOWNED_STATE = bind("combat.hiredVillagersUseDownedState", Boolean.class);
     public static final ConfigValue<Boolean> PARTY_VILLAGERS_USE_DOWNED_STATE = bind("combat.partyVillagersUseDownedState", Boolean.class);
+    public static final ConfigValue<Boolean> PLAYER_DAMAGE_DOWNS_ELIGIBLE_VILLAGERS = bind("combat.playerDamageDownsEligibleVillagers", Boolean.class);
+    public static final ConfigValue<Boolean> MOB_DAMAGE_DOWNS_ELIGIBLE_VILLAGERS = bind("combat.mobDamageDownsEligibleVillagers", Boolean.class);
+    public static final ConfigValue<Boolean> ENVIRONMENTAL_DAMAGE_DOWNS_ELIGIBLE_VILLAGERS = bind("combat.environmentalDamageDownsEligibleVillagers", Boolean.class);
     public static final ConfigValue<Integer> DOWNED_MINIMUM_TICKS = bind("combat.downedMinimumTicks", Integer.class);
     public static final ConfigValue<Double> DOWNED_RECOVERY_HEALTH_PERCENT = bind("combat.downedRecoveryHealthPercent", Double.class);
     public static final ConfigValue<Double> DOWNED_THREAT_RADIUS = bind("combat.downedThreatRadius", Double.class);